fix: Search bar in FollowerListVC
When deleting the search characters one by one, the last deletion will not trigger loading full array of followers.
We can fix it by calling updateData(on: followers) one the updateSearchResults is empty.
And since this fix also works when hitting the cancel button of search bar, we can just remove the UISearchBarDelegate and related methods.